# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix Online Booking Service API
  version: 1.0.0
  description: API contract for the Phoenix Online Booking Service

servers:
  - url: http://localhost:8083

tags:
  - name: Bookings
    description: Booking lifecycle operations

paths:
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a new booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
            examples:
              sample:
                value:
                  eventId: EVT-1001
                  customerEmail: christy@example.com
                  ticketType: VIP
                  quantity: 2
                  totalAmount: 5000.00
                  seat: A123
                  userId: usr-001
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Validation or business rule failure
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "502":
          description: Downstream integration error
    get:
      tags: [Bookings]
      summary: List all bookings
      description: Requires JWT permission VIEW_ALL_BOOKINGS
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All bookings returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking by bookingId
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          example: BKG-ABC1234567
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Booking not found
    patch:
      tags: [Bookings]
      summary: Update booking details
      description: Partial update. Requires JWT permission EDIT_BOOKING. Does not change booking or payment status.
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          example: BKG-ABC1234567
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
            examples:
              seatOnly:
                value:
                  seat: B200
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Validation or no fields to update
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Booking not found

  /bookings/customer/{email}:
    get:
      tags: [Bookings]
      summary: Get bookings by customer email
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
          example: christy@example.com
      responses:
        "200":
          description: Booking list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /bookings/{bookingId}/cancel:
    patch:
      tags: [Bookings]
      summary: Cancel booking
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          example: BKG-ABC1234567
      responses:
        "200":
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Invalid booking state
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Booking not found

  /bookings/{bookingId}/start-payment:
    post:
      tags: [Bookings]
      summary: Start payment for a booking
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          example: BKG-ABC1234567
      responses:
        "200":
          description: Payment session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartPaymentResponse"
        "400":
          description: Invalid booking state
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Booking not found
        "502":
          description: Downstream integration error

  /bookings/{bookingId}/expire:
    post:
      tags: [Bookings]
      summary: Expire booking
      security:
        - internalServiceToken: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          example: BKG-ABC1234567
      responses:
        "200":
          description: Booking expired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Invalid booking state
        "401":
          description: Unauthorized
        "404":
          description: Booking not found

  /bookings/payment-callback:
    post:
      tags: [Bookings]
      summary: Handle payment callback
      security:
        - internalServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCallbackRequest"
            examples:
              success:
                value:
                  bookingId: BKG-ABC1234567
                  paymentReferenceId: PAY-123456
                  paymentStatus: SUCCESS
                  transactionId: TXN-0001
      responses:
        "200":
          description: Payment callback processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Invalid callback payload or state
        "401":
          description: Unauthorized
        "404":
          description: Booking not found
        "502":
          description: Downstream integration error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token

  schemas:
    UpdateBookingRequest:
      type: object
      description: All properties optional; at least one must be provided.
      properties:
        eventId:
          type: string
          example: EVT-1001
        customerEmail:
          type: string
          format: email
          example: christy@example.com
        seat:
          type: string
          example: A123
        userId:
          type: string
          example: usr-001
        ticketType:
          type: string
          example: VIP
        quantity:
          type: integer
          minimum: 1
          example: 2
        totalAmount:
          type: number
          format: double
          minimum: 0.01
          example: 5000.00

    CreateBookingRequest:
      type: object
      required:
        - eventId
        - customerEmail
        - ticketType
        - quantity
        - totalAmount
        - seat
        - userId
      properties:
        eventId:
          type: string
          example: EVT-1001
        customerEmail:
          type: string
          example: christy@example.com
        ticketType:
          type: string
          example: VIP
        quantity:
          type: integer
          example: 2
        totalAmount:
          type: number
          format: double
          example: 5000.00
        seat:
          type: string
          example: A123
        userId:
          type: string
          example: usr-001

    PaymentCallbackRequest:
      type: object
      required:
        - bookingId
        - paymentReferenceId
        - paymentStatus
      properties:
        bookingId:
          type: string
          example: BKG-ABC1234567
        paymentReferenceId:
          type: string
          example: PAY-123456
        paymentStatus:
          type: string
          example: SUCCESS
        transactionId:
          type: string
          example: TXN-0001

    StartPaymentResponse:
      type: object
      properties:
        bookingId:
          type: string
          example: BKG-ABC1234567
        paymentReferenceId:
          type: string
          example: PAY-123456
        bookingStatus:
          type: string
          example: AWAITING_PAYMENT
        paymentStatus:
          type: string
          example: PENDING

    BookingResponse:
      type: object
      properties:
        id:
          type: string
          example: 67f8bc9d1234567890abcdef
        bookingId:
          type: string
          example: BKG-ABC1234567
        eventId:
          type: string
          example: EVT-1001
        userId:
          type: string
          example: usr-001
        customerEmail:
          type: string
          example: christy@example.com
        seat:
          type: string
          example: A123
        ticketType:
          type: string
          example: VIP
        quantity:
          type: integer
          example: 2
        totalAmount:
          type: number
          format: double
          example: 5000.00
        paymentReferenceId:
          type: string
          example: PAY-123456
        paymentTransactionId:
          type: string
          example: TXN-0001
        bookingStatus:
          type: string
          example: CONFIRMED
        paymentStatus:
          type: string
          example: SUCCESS
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time